requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.99.1
fastapi-jwt-auth==0.5.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
MarkupSafe==2.1.3
psycopg2==2.9.7
psycopg2-binary==2.9.7
pydantic===1.10.11
pydantic_core==2.6.3
PyJWT==1.7.1
sniffio==1.3.0
SQLAlchemy==2.0.21
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
Werkzeug==2.3.7

python-dotenv~=1.0.0